// src/squads/user/common/types/user-group.ts
import { ArrayElement } from "src/common/constants/types";
import {
    User_LocationListByIdsQuery,
    User_GrantedRoleAccessPathByGrantedRoleIdsQuery,
    User_UserGroupOneQuery,
} from "src/squads/user/service/bob/bob-types";

import { LocationInformation } from "./common";

export interface Role {
    role_id: string;
    role_name: string;
}
export type UserGroup = ArrayElement<User_UserGroupOneQuery["user_group"]>;
export interface UserGroupPackage {
    role?: Role;
    user_group_id?: string;
    user_group_name: string;
    locations?: User_LocationListByIdsQuery["locations"];
    isDraft?: boolean;
    id?: string;
}

export type GrantedRoleAccessPath = ArrayElement<
    User_GrantedRoleAccessPathByGrantedRoleIdsQuery["granted_role_access_path"]
>;
export interface GrantedPermission {
    id?: string; // generated by field array of react-hook-form
    granted_role_id: string;
    role: Role;
    locations: LocationInformation[];
}

export interface UserGroupInfo {
    id?: string;
    name: string;
    grantedPermissionPackage: GrantedPermission[];
}